requirements.txt
flask>=3.0
pyyaml>=6.0
torch>=2.0
soundfile>=0.12
numpy>=1.24
kokoro>=0.7
